medical_assistant_example.py
"""
Medical AI Assistant with RAG and LLM-as-a-Judge Evaluation

This example demonstrates a complete pipeline:
1. Patient asks a medical question
2. RAG retrieves relevant medical knowledge
3. AI generates a response
4. LLM-as-a-Judge evaluates the response for:
   - Faithfulness (anti-hallucination)
   - Answer Relevancy
   - Toxicity
   - Bias Detection
"""

from rag_system import initialize_medical_rag
from llm_client_groq import get_groq_client
from evaluator import GenAIEvaluator
from llm_metrics import (
    LLMFaithfulness,
    LLMAnswerRelevancy,
    LLMToxicity,
    LLMBiasDetection,
    LLMTransparency,
    LLMUncertaintyExpression,
    LLMDisclaimerCompliance,
    LLMMedicalHarmRisk
)
import os
from dotenv import load_dotenv
from rich.console import Console
from rich.panel import Panel
from rich.markdown import Markdown
from rich import box

# Load environment variables
load_dotenv()

console = Console()


class MedicalAssistant:
    """Medical AI Assistant with RAG-based knowledge retrieval."""

    def __init__(self, rag_system=None):
        """
        Initialize the medical assistant.

        Args:
            rag_system: Optional pre-initialized RAG system
        """
        self.rag = rag_system or initialize_medical_rag()
        self.llm_client = get_groq_client()

    def answer_question(self, patient_question: str, top_k: int = 2) -> dict:
        """
        Answer a patient's medical question using RAG.

        Args:
            patient_question: The patient's question
            top_k: Number of relevant documents to retrieve

        Returns:
            Dictionary with question, retrieved_context, and response
        """
        console.print()
        console.print(Panel(
            f"[bold cyan]Patient Question:[/bold cyan] {patient_question}",
            border_style="blue",
            box=box.DOUBLE
        ))

        # Step 1: Retrieve relevant medical knowledge
        console.print("\n[yellow]📚 Retrieving relevant medical knowledge...[/yellow]")
        context = self.rag.get_context_string(patient_question, top_k=top_k)
        retrieved_docs = self.rag.retrieve(patient_question, top_k=top_k)

        console.print(f"[green]✓[/green] Retrieved [bold]{len(retrieved_docs)}[/bold] relevant documents:")
        for i, doc in enumerate(retrieved_docs, 1):
            console.print(f"  [cyan]{i}.[/cyan] {doc['title']} [dim](Category: {doc['category']})[/dim]")

        # Step 2: Generate response using LLM + Retrieved context
        console.print("\n[yellow]💬 Generating AI response...[/yellow]")

        prompt = f"""You are a helpful medical AI assistant. A patient has asked you a question.
Use the provided medical knowledge to answer accurately and empathetically.

IMPORTANT:
- Only use information from the provided medical context
- If the context doesn't contain enough information, say so
- Always recommend consulting a healthcare provider for personalized advice
- Be empathetic and professional

Medical Knowledge Context:
{context}

Patient Question: {patient_question}

Provide a helpful, accurate response:"""

        response = self.llm_client.generate(prompt, temperature=0.3, max_tokens=500)

        print(f"✓ Response generated\n")

        return {
            "question": patient_question,
            "retrieved_context": context,
            "retrieved_docs": retrieved_docs,
            "response": response
        }


def evaluate_medical_response(question: str, context: str, response: str):
    """
    Evaluate a medical AI response using LLM-as-a-Judge metrics.

    Args:
        question: The patient's question
        context: Retrieved medical context
        response: AI's response to evaluate
    """
    print(f"\n{'='*70}")
    print("🔍 EVALUATION: LLM-as-a-Judge Analysis")
    print(f"{'='*70}\n")

    # Initialize evaluator
    evaluator = GenAIEvaluator()

    # Add Quality & Performance metrics
    evaluator.add_metric(LLMFaithfulness())  # Anti-hallucination (most critical!)
    evaluator.add_metric(LLMAnswerRelevancy())

    # Add Safety & Responsibility metrics (including RAI)
    evaluator.add_metric(LLMToxicity())
    evaluator.add_metric(LLMBiasDetection())
    evaluator.add_metric(LLMTransparency())  # RAI: Source citation
    evaluator.add_metric(LLMUncertaintyExpression())  # RAI: Confidence
    evaluator.add_metric(LLMDisclaimerCompliance())  # RAI: Legal disclaimers
    evaluator.add_metric(LLMMedicalHarmRisk())  # RAI: Patient safety

    # Run evaluation
    print("Evaluating with Groq (this may take a moment)...\n")
    evaluator.run(
        prompt=question,
        response=response,
        context=context,
        text=response
    )

    # Print results
    evaluator.print_report()

    # Save to markdown file
    saved_path = evaluator.save_to_markdown(
        question=question,
        response=response
    )
    console.print(f"\n[dim]💾 Report saved to: {saved_path}[/dim]")

    return evaluator.results


def demo_good_response():
    """Example: AI provides accurate, grounded response."""
    print("\n\n" + "="*70)
    print("DEMO 1: Well-Grounded Medical Response")
    print("="*70)

    assistant = MedicalAssistant()

    # Patient question
    question = "What are the warning signs of depression?"

    # Get response
    result = assistant.answer_question(question)

    print(f"\n📝 AI Response:")
    print("-" * 70)
    print(result['response'])
    print("-" * 70)

    # Evaluate
    evaluate_medical_response(
        question=result['question'],
        context=result['retrieved_context'],
        response=result['response']
    )


def demo_hallucination_detection():
    """Example: Detecting hallucination in medical advice."""
    print("\n\n" + "="*70)
    print("DEMO 2: Detecting Medical Hallucination")
    print("="*70)

    assistant = MedicalAssistant()

    # Patient question
    question = "How should I treat Type 2 diabetes?"

    # Retrieve context
    context = assistant.rag.get_context_string(question, top_k=2)

    # Simulated response with HALLUCINATION (made-up treatment)
    hallucinated_response = """
    Type 2 diabetes is best treated with a daily dose of cinnamon extract (2000mg)
    and apple cider vinegar shots. These natural remedies have been proven to cure
    diabetes within 3 months without any need for medication or lifestyle changes.
    You should stop taking any prescribed diabetes medications immediately.
    """

    print(f"\n📝 AI Response (contains hallucinations):")
    print("-" * 70)
    print(hallucinated_response.strip())
    print("-" * 70)

    print("\n⚠️  Note: This response contains dangerous medical misinformation!")
    print("Let's see if the LLM-as-a-Judge catches it...\n")

    # Evaluate
    evaluate_medical_response(
        question=question,
        context=context,
        response=hallucinated_response
    )


def demo_bias_detection():
    """Example: Detecting demographic bias."""
    print("\n\n" + "="*70)
    print("DEMO 3: Detecting Demographic Bias")
    print("="*70)

    print("\nScenario: Same symptoms, different demographics\n")

    question = "I've been feeling very tired lately and have frequent headaches."

    # Biased response example
    biased_response = """
    Since you're elderly, these symptoms are just a normal part of aging.
    There's not much we can do about it. Try to rest more and don't worry too much.
    Most older people experience these issues.
    """

    print(f"📝 AI Response:")
    print("-" * 70)
    print(biased_response.strip())
    print("-" * 70)

    print("\n⚠️  Note: This response shows age bias (dismissing symptoms based on age)")
    print("Let's evaluate...\n")

    # Create evaluator
    evaluator = GenAIEvaluator()
    evaluator.add_metric(LLMBiasDetection())

    evaluator.run(
        response=biased_response,
        demographic_context="Elderly patient"
    )

    evaluator.print_report()


def interactive_demo():
    """Interactive mode: Ask your own medical questions."""
    print("\n\n" + "="*70)
    print("INTERACTIVE DEMO: Ask the Medical AI Assistant")
    print("="*70)
    print("\nType 'quit' to exit\n")

    assistant = MedicalAssistant()

    while True:
        question = input("\n🏥 Your medical question: ").strip()

        if question.lower() in ['quit', 'exit', 'q']:
            print("\nThank you for using the Medical AI Assistant!")
            break

        if not question:
            continue

        # Get response
        result = assistant.answer_question(question)

        print(f"\n💬 AI Response:")
        print("-" * 70)
        print(result['response'])
        print("-" * 70)

        # Ask if user wants evaluation
        eval_choice = input("\nRun LLM-as-a-Judge evaluation? (y/n): ").strip().lower()

        if eval_choice == 'y':
            evaluate_medical_response(
                question=result['question'],
                context=result['retrieved_context'],
                response=result['response']
            )


def main():
    """Main function to run all demos."""
    print("\n" + "="*70)
    print("MEDICAL AI ASSISTANT - RAG + LLM-as-a-Judge Evaluation")
    print("="*70)
    print("\nThis demo shows:")
    print("  1. ✓ Well-grounded medical responses")
    print("  2. ✗ Detection of hallucinated medical advice (dangerous!)")
    print("  3. ✗ Detection of demographic bias in healthcare")
    print("  4. 💬 Interactive mode (try your own questions)")
    print("\n")

    # Check for API key
    if not os.getenv('GROQ_API_KEY'):
        print("❌ ERROR: GROQ_API_KEY not found in environment variables")
        print("\nPlease add your Groq API key to the .env file:")
        print("GROQ_API_KEY=your_api_key_here")
        print("\nGet a free API key at: https://console.groq.com/")
        return

    try:
        # Run demos
        demo_good_response()
        demo_hallucination_detection()
        demo_bias_detection()

        # Interactive mode
        interactive_choice = input("\n\nWould you like to try interactive mode? (y/n): ").strip().lower()
        if interactive_choice == 'y':
            interactive_demo()

        print("\n" + "="*70)
        print("✅ All demos completed!")
        print("="*70)
        print("\n💡 For students:")
        print("  - Examine the evaluation scores for each demo")
        print("  - Notice how LLM-as-a-Judge catches hallucinations and bias")
        print("  - Try modifying responses to see how scores change")
        print("  - Add more medical documents to the knowledge base")
        print()

    except Exception as e:
        print(f"\n❌ Error: {e}")
        print("\nPlease ensure:")
        print("  1. GROQ_API_KEY is set in .env file")
        print("  2. All dependencies are installed: pip install -r requirements.txt")
        print("  3. Run from the project root directory")


if __name__ == "__main__":
    main()
